requirements.txt
streamlit==1.38.0
pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
opencv-python-headless==4.10.0.84
streamlit-drawable-canvas==0.9.3
